% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOutbreak.R
\name{simOutbreak}
\alias{simOutbreak}
\title{Branching Process Outbreak Simulator}
\usage{
simOutbreak(
  timeImport,
  R = 1.97,
  k = 1000,
  si_mean = 4.56,
  si_sd = 0.95,
  inc_shape = 1.88,
  inc_scale = 7.97,
  symp_to_death_mean = 22.3,
  symp_to_death_sd = 0.42 * 22.3,
  report_delay_rate = 3,
  stopSimulationDay = 68,
  repSims = 100,
  asympProp = 0.35,
  asympRFraction = 0.5,
  lnormFlag = TRUE
)
}
\arguments{
\item{timeImport}{timing of each imported infection}

\item{R}{reproduction number}

\item{k}{dispersion parameter}

\item{si_mean}{mean of serial interval distribution}

\item{si_sd}{standard deviation of serial interval distribution}

\item{inc_shape}{hape parameter of incubation period distribution (Weibull)}

\item{inc_scale}{scale parameter of incubation period distribution (Weibull)}

\item{symp_to_death_mean}{mean of time between symptom onset and death}

\item{symp_to_death_sd}{std. dev. of time between symptom onset and death}

\item{report_delay_rate}{mean delay between symptoms and reporting}

\item{stopSimulationDay}{day of year since Jan 1 when simulation stops}

\item{repSims}{number of replicates to produce for these parameters}

\item{asympProp}{proportion of infections that are asymptomatic}

\item{lnormFlag}{toggles whether serial interval distribution is lognormal}

\item{asympRFractionrelative}{infectiousness of asymptomatics}
}
\description{
This function works exactly the same as the R version.
If both are loaded in the same R session, use this one via \code{simOutbreak::simOutbreak}
}
