% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOutbreak.R
\name{simOutbreak}
\alias{simOutbreak}
\title{Branching Process Outbreak Simulator}
\usage{
simOutbreak(
  timeImport,
  R = 1.97,
  k = 1000,
  si_mean = 4.56,
  si_sd = 0.95,
  inc_shape = 1.88,
  inc_scale = 7.97,
  symp_to_death_mean = 22.3,
  symp_to_death_sd = 0.42 * 22.3,
  report_delay_rate = 3,
  stopSimulationDay = 68,
  repSims = 100,
  asympProp = 0.35,
  asympRFraction = 0.5,
  lnormFlag = TRUE
)
}
\description{
Branching Process Outbreak Simulator
}
